Decouple LeonardoAzureSuite from LeonardoSuite
